+++ src/particleGen/specific/includes/gammaPolarised.hh
@@ -1,8 +1,8 @@
 
 //
 //
-//    Copyright (C) 2019-2022 Universitat de València - UV
-//    Copyright (C) 2019-2022 Universitat Politècnica de València - UPV
+//    Copyright (C) 2019-2023 Universitat de València - UV
+//    Copyright (C) 2019-2023 Universitat Politècnica de València - UPV
 //
 //    This file is part of PenRed: Parallel Engine for Radiation Energy Deposition.
 //
@@ -54,6 +54,7 @@ class gammaPolarised_specificSampler : public abc_specificSampler<pen_state_gPol
   
   int configure(double& /*Emax*/,
 		const pen_parserSection& config,
+		const unsigned /*nthreads*/,
 		const unsigned verbose);
 };