+++ src/particleGen/specific/source/CTsource.cpp
@@ -1,8 +1,8 @@
 
 //
 //
-//    Copyright (C) 2020-2021 Universitat de València - UV
-//    Copyright (C) 2020-2021 Universitat Politècnica de València - UPV
+//    Copyright (C) 2020-2023 Universitat de València - UV
+//    Copyright (C) 2020-2023 Universitat Politècnica de València - UPV
 //
 //    This file is part of PenRed: Parallel Engine for Radiation Energy Deposition.
 //
@@ -37,6 +37,7 @@ void ct_specificSampler::skip(const unsigned long long dhists){
 
 int ct_specificSampler::configure(double& Emax,
 				  const pen_parserSection& config,
+				  const unsigned nthreads,
 				  const unsigned verbose){
   
   //First, initialize the phase space file sampler
@@ -111,7 +112,7 @@ int ct_specificSampler::configure(double& Emax,
       return err;
     }
     
-    err = psf.configure(Emax,psfSection,verbose);
+    err = psf.configure(Emax,psfSection,nthreads,verbose);
     if(err != 0){
       if(verbose > 0)
 	printf("ctSource:configure: Error: Unable to configure psf.\n"