+++ src/particleGen/specific/source/gammaPolarised.cpp
@@ -1,8 +1,8 @@
 
 //
 //
-//    Copyright (C) 2019-2022 Universitat de València - UV
-//    Copyright (C) 2019-2022 Universitat Politècnica de València - UPV
+//    Copyright (C) 2019-2023 Universitat de València - UV
+//    Copyright (C) 2019-2023 Universitat Politècnica de València - UPV
 //
 //    This file is part of PenRed: Parallel Engine for Radiation Energy Deposition.
 //
@@ -41,6 +41,7 @@ void gammaPolarised_specificSampler::sample(pen_state_gPol& state,
 
 int gammaPolarised_specificSampler::configure(double& /*Emax*/,
 					      const pen_parserSection& config,
+					      const unsigned /*nthreads*/,
 					      const unsigned verbose){
 
   int err = 0;